Update `Settings` struct to use `u32` for dimensions
Changed all dimension fields in the `Settings` struct from `u16` to `u32` to support larger values. This ensures consistency and allows for greater flexibility in handling settings-related data.

+++ src/settings/mod.rs
@@ -26,14 +26,14 @@ pub fn global_settings() -> &'static Mutex<Settings> {
 
 #[derive(Serialize, Deserialize, Debug, Default)]
 pub struct Settings {
-    pub(crate) ocr_x: u16,
-    pub(crate) ocr_y: u16,
-    pub(crate) ocr_width: u16,
-    pub(crate) ocr_height: u16,
-    pub(crate) icon_x: u16,
-    pub(crate) icon_y: u16,
-    pub(crate) icon_width: u16,
-    pub(crate) icon_height: u16,
+    pub(crate) ocr_x: u32,
+    pub(crate) ocr_y: u32,
+    pub(crate) ocr_width: u32,
+    pub(crate) ocr_height: u32,
+    pub(crate) icon_x: u32,
+    pub(crate) icon_y: u32,
+    pub(crate) icon_width: u32,
+    pub(crate) icon_height: u32,
 }
 
 impl Settings {
